Refactor: Progress service
